refactor: 예약 서비스 코드 책임분리
