+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -13,17 +13,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.reservation.ReservationResponses;
 import roomescape.infrastructure.LoginUserId;
-import roomescape.service.ReservationService;
+import roomescape.service.AdminReservationService;
 
 @Validated
 @RestController
 @RequestMapping("/admin/reservations")
 public class AdminReservationController {
+    private final AdminReservationService adminReservationService;
 
-    private final ReservationService reservationService;
-
-    public AdminReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
+    public AdminReservationController(AdminReservationService adminReservationService) {
+        this.adminReservationService = adminReservationService;
     }
 
     @GetMapping
@@ -33,12 +32,12 @@ public class AdminReservationController {
             @RequestParam(defaultValue = "20") @Min(1) @Max(100) int size,
             @RequestParam(required = false) @Size(min = 1) String name
     ) {
-        return ResponseEntity.ok(reservationService.getReservations(page, size, name, managerId));
+        return ResponseEntity.ok(adminReservationService.getReservations(page, size, name, managerId));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@LoginUserId Long managerId, @PathVariable Long id) {
-        reservationService.deleteReservation(id, managerId);
+        adminReservationService.deleteReservation(id, managerId);
         return ResponseEntity.ok().build();
     }
 }
